+++ src/GraphBuilding/Structures/Graph.cpp
@@ -34,7 +34,7 @@ void Graph::addEdge(unsigned int from, unsigned int to, dist_t weight) {
 
 //______________________________________________________________________________________________________________________
 unsigned int Graph::nodes()const {
-    return this->followingNodes.size();
+    return (unsigned int) this->followingNodes.size();
 }
 
 //______________________________________________________________________________________________________________________
